+++ FreePlayTraining/GameInformation.h
@@ -6,5 +6,5 @@ struct GameInformation {
 	BallWrapper Ball;
 	CarWrapper Car;
 	std::shared_ptr<CVarManagerWrapper> CVarManager;
-	float DeltaTime;
+	double DeltaTime;
 };
